32 bits

+++ bs/pack.h
@@ -1,5 +1,5 @@
 #include <stdint.h>
 #include <string.h>
 
-void pack(uint16_t *out, const uint8_t *in);
-void unpack(uint8_t *out, const uint16_t *in);
+void pack(uint32_t *out, const uint8_t *in);
+void unpack(uint8_t *out, const uint32_t *in);
